GO enrichment: more member functions added for the result object: 'oddsRatios', 'expectedCounts', 'geneIdUniverse', 'condGeneIdUniverse', geneIdsByCategory', 'sigCategories'.
Also, we return the category -> Entrez mapping now as part of the
result, probably this is not too big, because we have it only once for
the whole gene list. And it is needed for some standard members, like
'geneIdUniverse'.

The name of the test was updated to a more informative one.